% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{AIC.lgc}
\alias{AIC.lgc}
\title{AIC for lgc Model}
\usage{
\method{AIC}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Numeric. AIC value.
}
\description{
Returns the AIC value for a fitted \code{lgc} model.
}
