% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{BIC.lgc}
\alias{BIC.lgc}
\title{BIC for lgc Model}
\usage{
\method{BIC}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Numeric. BIC value.
}
\description{
Returns the BIC value for a fitted \code{lgc} model.
}
