% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{coefficients.lgc}
\alias{coefficients.lgc}
\title{Coefficients for lgc Model}
\usage{
\method{coefficients}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Matrix of parameter estimates.
}
\description{
Returns the estimated coefficients from a fitted \code{lgc} model.
}
