% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{ComputeLimits}
\alias{ComputeLimits}
\title{Compute Truncation Limits for Latent Gaussian Count Models}
\usage{
ComputeLimits(mod, Parms, t, Zhat, Rt)
}
\arguments{
\item{mod}{A list containing model specification and data, typically created by \code{\link{ModelScheme}}.}

\item{Parms}{A list of model parameters retrieved via \code{\link{RetrieveParameters}}.}

\item{t}{Integer. Current time point.}

\item{Zhat}{Numeric vector of predicted latent values up to time \code{t}.}

\item{Rt}{Numeric vector of innovation standard deviations from the innovations algorithm.}
}
\value{
A list with elements \code{a} and \code{b} representing the lower and upper limits
for truncated normal sampling.
}
\description{
Computes the lower and upper truncation limits used in the sampling of the latent variable
in the particle filter algorithm. Corresponds to equation (19) in the JASA paper.
}
\seealso{
\code{\link{SampleTruncNormParticles}}
}
\keyword{internal}
