% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{ComputeZhat_t}
\alias{ComputeZhat_t}
\title{Compute Filtered Latent Value at Time t}
\usage{
ComputeZhat_t(mod, IA, Z, Zhat, t, Parms)
}
\arguments{
\item{mod}{Model list as returned by \code{\link{ModelScheme}}.}

\item{IA}{A list returned by \code{\link{InnovAlg}}, containing innovation variances and
theta coefficients.}

\item{Z}{Matrix of latent particle values.}

\item{Zhat}{Matrix of predicted latent means.}

\item{t}{Integer. Time point at which to compute the latent mean.}

\item{Parms}{List of model parameters from \code{\link{RetrieveParameters}}.}
}
\value{
A numeric vector with the filtered latent value at time \code{t}.
}
\description{
Computes the conditional mean of the latent variable at time \code{t} using the innovations
algorithm output.
}
\keyword{internal}
