% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\docType{package}
\name{countsFun-package}
\alias{countsFun}
\alias{countsFun-package}
\title{countsFun: Fit Count Time Series Models}
\description{
The lgc package fits Latent Gaussian Count time series models using copula transformations and simulation inference techniques developed by Jia et al. (2023).
}
\author{
\strong{Maintainer}: Stefanos Kechagias \email{stefanoskeh@gmail.com}

Authors:
\itemize{
  \item James Livsey \email{james.a.livsey@gmail.com}
  \item Robert Lund \email{lund@clemson.edu}
  \item Vladas Pipiras \email{pipiras@email.unc.edu}
}

}
