% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{DIM}
\alias{DIM}
\title{Dimension function for vector/matrix inputs
Treats vectors as length x 1 column vectors}
\usage{
DIM(x)
}
\arguments{
\item{x}{vector (column matrix) or matrix}
}
\value{
the vector's dimension
}
\description{
Dimension function for vector/matrix inputs
Treats vectors as length x 1 column vectors
}
\keyword{internal}
