% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{dmixpois1}
\alias{dmixpois1}
\title{Density of the Mixed Poisson Distribution}
\usage{
dmixpois1(x, lam1, lam2, p)
}
\arguments{
\item{x}{Integer or numeric vector. The value(s) at which to evaluate the PMF.}

\item{lam1}{Numeric. Mean of the first Poisson component.}

\item{lam2}{Numeric. Mean of the second Poisson component.}

\item{p}{Numeric between 0 and 1. Mixing probability for the first component.}
}
\value{
A numeric vector of probabilities corresponding to \code{x}.
}
\description{
Computes the probability mass function (PMF) of a two-component mixed Poisson distribution.
}
\details{
The mixed Poisson PMF is given by:
\deqn{p * dpois(x, lam1) + (1 - p) * dpois(x, lam2)}
}
\examples{
dmixpois1(0:10, lam1 = 2, lam2 = 5, p = 0.6)

}
