% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{get_rand_state}
\alias{get_rand_state}
\title{Get current random seed state
Add some functions that I will need for common random numbers}
\usage{
get_rand_state()
}
\description{
Get current random seed state
Add some functions that I will need for common random numbers
}
\keyword{internal}
