% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{logLik.lgc}
\alias{logLik.lgc}
\title{Log-Likelihood for lgc Model}
\usage{
\method{logLik}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Numeric. Log-likelihood value.
}
\description{
Returns the log-likelihood value stored in a fitted \code{lgc} model object.
}
