% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_hitters}
\alias{no_hitters}
\title{Major League Baseball No-Hitters}
\format{
A data frame with \code{n} observations and 6 variables:
\describe{
\item{Date}{Character. The date of the game (MM/DD/YYYY format).}
\item{Pitcher}{Character. Name of the pitcher who threw the no-hitter.}
\item{Team}{Character. Team for which the pitcher played.}
\item{Lg}{Character. League of the team (e.g., NA, NL, AL). May contain missing values.}
\item{Plate.Umpire}{Character. Name of the home plate umpire.}
\item{Notes}{Character. Additional notes (e.g., "Perfect Gm"). May be empty.}
}
}
\source{
Retrosheet, Inc. \url{http://www.retrosheet.org}
}
\usage{
no_hitters
}
\description{
A dataset containing a list of Major League Baseball no-hitter games,
compiled from Retrosheet, Inc. (\url{http://www.retrosheet.org}).
}
\details{
The \code{no_hitters} dataset provides historical records of no-hitter
games in Major League Baseball. A no-hitter is a game in which a pitcher
(or multiple pitchers) allows no hits over the course of a complete game.

The data are sourced from Retrosheet and include early baseball history,
with records dating back to the 19th century.
}
\examples{
data(no_hitters)
head(no_hitters)

}
\keyword{datasets}
