% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{nobs.lgc}
\alias{nobs.lgc}
\title{Number of Observations in lgc Model}
\usage{
\method{nobs}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An integer, the number of observations.
}
\description{
Returns the number of observations used in fitting the model.
}
