% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{pgpois1}
\alias{pgpois1}
\title{Generalized Poisson CDF (single mean)}
\usage{
pgpois1(x, a, m)
}
\arguments{
\item{x}{Integer or numeric vector. Values at which to evaluate the CDF.}

\item{a}{Numeric. Dispersion parameter.}

\item{m}{Numeric. Mean parameter.}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Computes the cumulative distribution function (CDF) of the Generalized Poisson distribution
for a single value of \code{m}.
}
\seealso{
\code{\link{pGpois}}, \code{\link{dGpois}}
}
