% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{pit_plot}
\alias{pit_plot}
\alias{pit_plot.lgc}
\title{PIT Histogram for \code{lgc} Objects}
\usage{
pit_plot(object, ...)

\method{pit_plot}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"lgc"}.}

\item{...}{Additional graphical arguments passed to \code{\link{plot}}.}
}
\value{
Invisibly returns a data frame containing:
\describe{
\item{mid}{Bin midpoints.}
\item{width}{Bin widths.}
\item{observed}{Observed PIT relative frequencies.}
\item{expected}{Expected uniform frequencies.}
\item{confint_lwr}{Lower confidence interval bound.}
\item{confint_upr}{Upper confidence interval bound.}
}
}
\description{
Produces a Probability Integral Transform (PIT) histogram for a fitted
\code{lgc} model object. The histogram includes:
\itemize{
\item observed PIT relative frequencies,
\item the expected uniform reference line,
\item binomial confidence interval bands.
}
}
\details{
The PIT histogram is used to assess probabilistic calibration of the
predictive distribution. Under correct calibration, the PIT histogram
should be approximately uniform.
}
\examples{
\dontrun{
CountDist = "Negative Binomial"
ARMAModel = c(3,0)
RegModel  = MOVE ~ 1 + Buy

data(drinksales)
df <- drinksales[1:104, ]

fit = lgc(
   RegModel = RegModel,
         df = df,
  ARMAModel = ARMAModel,
  CountDist = CountDist,
       Task = "Optimization"
)

pit_plot(fit)
}

}
