% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{plot.lgc}
\alias{plot.lgc}
\title{Diagnostic Plots for lgc Objects}
\usage{
\method{plot}{lgc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"lgc"}.}

\item{...}{Additional graphical arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a 2 by 2 panel of diagnostic plots for a fitted \code{lgc} object:
PIT histogram, normal QQ plot of residuals, ACF of residuals, and PACF of
residuals.
}
