% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{pmixpois1}
\alias{pmixpois1}
\title{Cumulative Distribution Function of the Mixed Poisson Distribution}
\usage{
pmixpois1(x, lam1, lam2, p)
}
\arguments{
\item{x}{Integer or numeric vector. The value(s) at which to evaluate the CDF.}

\item{lam1}{Numeric. Mean of the first Poisson component.}

\item{lam2}{Numeric. Mean of the second Poisson component.}

\item{p}{Numeric between 0 and 1. Mixing probability for the first component.}
}
\value{
A numeric vector of cumulative probabilities corresponding to \code{x}.
}
\description{
Computes the cumulative distribution function (CDF) of a two-component mixed Poisson distribution.
}
\details{
The mixed Poisson CDF is given by:
\deqn{p * ppois(x, lam1) + (1 - p) * ppois(x, lam2)}
}
\examples{
pmixpois1(0:10, lam1 = 2, lam2 = 5, p = 0.6)

}
