% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{print.lgc}
\alias{print.lgc}
\title{Print an lgc Model}
\usage{
\method{print}{lgc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Nicely formats and prints a fitted \code{lgc} model object.
}
\examples{
\dontrun{
  print(mylgc)
}
}
