% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{qGpois}
\alias{qGpois}
\title{Generalized Poisson Quantile Function (vectorized and scalar)}
\usage{
qGpois(p, a, m)
}
\arguments{
\item{p}{Numeric vector of probabilities (between 0 and 1).}

\item{a}{Numeric. Dispersion parameter.}

\item{m}{Numeric scalar or vector. Mean parameter(s).}
}
\value{
A numeric vector or matrix of quantiles.
}
\description{
Computes the quantile function (inverse CDF) of the Generalized Poisson distribution
for one or more values of the mean parameter \code{m}.
}
\examples{
qGpois(0.8,0.4,3)

}
\seealso{
\code{\link{pGpois}}
}
