% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{qgpois1}
\alias{qgpois1}
\title{Generalized Poisson Quantile Function (single mean)}
\usage{
qgpois1(p, a, m)
}
\arguments{
\item{p}{Numeric vector of probabilities (between 0 and 1).}

\item{a}{Numeric. Dispersion parameter.}

\item{m}{Numeric. Mean parameter.}
}
\value{
A numeric vector of quantiles.
}
\description{
Computes the quantile function (inverse CDF) of the Generalized Poisson distribution
for a single value of the mean.
}
\seealso{
\code{\link{qGpois}}, \code{\link{pGpois}}
}
