% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{qmixpois}
\alias{qmixpois}
\title{Vectorized Quantile Function for Mixed Poisson Distribution}
\usage{
qmixpois(p, lam1, lam2, prob)
}
\arguments{
\item{p}{Numeric vector of probabilities (between 0 and 1).}

\item{lam1}{Numeric or vector. Mean(s) of the first Poisson component.}

\item{lam2}{Numeric or vector. Mean(s) of the second Poisson component.}

\item{prob}{Numeric between 0 and 1. Mixing probability for the first component.}
}
\value{
A numeric vector of quantiles corresponding to \code{p}.
}
\description{
Computes quantiles of the mixed Poisson distribution using a vectorized approach
with \code{mapply()}.
}
\details{
This function is a vectorized version of \code{\link{qmixpois1}} using \code{mapply()}
to compute quantiles efficiently.
}
\examples{
qmixpois(c(0.1, 0.5, 0.9), lam1 = 2, lam2 = 5, prob = 0.6)

}
