% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{ResampleParticles}
\alias{ResampleParticles}
\title{Resample Particles Using Effective Sample Size (ESS) Criterion}
\usage{
ResampleParticles(mod, wgh, t, Znew)
}
\arguments{
\item{mod}{Model list as returned by \code{\link{ModelScheme}}.}

\item{wgh}{Matrix of particle weights for all time points.}

\item{t}{Integer. Current time index.}

\item{Znew}{Vector of new particle samples to be resampled.}
}
\value{
A resampled vector of particles of length equal to \code{mod$ParticleNumber}.
}
\description{
Performs resampling of particles when the effective sample size falls below a
threshold defined by the model. Follows relation (26) in the JASA paper.
}
\keyword{internal}
