% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{residuals.lgc}
\alias{residuals.lgc}
\title{Extract Residuals from an lgc Model}
\usage{
\method{residuals}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of residuals of length \code{nobs(object)}.
}
\description{
Computes residuals from a fitted \code{lgc} model using the
ARMA structure specified in the model.
}
\examples{
\dontrun{
  residuals(mylgc)
}
}
