% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{rGpois}
\alias{rGpois}
\title{Random Generation from Generalized Poisson Distribution}
\usage{
rGpois(n, a, m)
}
\arguments{
\item{n}{Integer. Number of observations to generate.}

\item{a}{Numeric. Dispersion parameter.}

\item{m}{Numeric or vector. Mean parameter(s).}
}
\value{
A numeric vector of random values drawn from the Generalized Poisson distribution.
}
\description{
Generates random observations from the Generalized Poisson distribution via inverse transform sampling.
}
\examples{
rGpois(10, a = 0.1, m = 3)

}
\seealso{
\code{\link{qGpois}}, \code{\link{dGpois}}
}
