% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{se.lgc}
\alias{se.lgc}
\title{Standard Errors for lgc Model}
\usage{
\method{se}{lgc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lgc}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Matrix of standard errors.
}
\description{
Returns the standard errors from a fitted \code{lgc} model.
}
