% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimPois2_AR1_075_N1k}
\alias{SimPois2_AR1_075_N1k}
\title{Simulated Poisson AR(1) Time Series}
\format{
A numeric vector of length 1000.
}
\source{
Simulated for package testing.
}
\usage{
SimPois2_AR1_075_N1k
}
\description{
A simulated Poisson time series with AR(1) dependence and phi = 0.75.
}
\keyword{datasets}
