% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{ComputeResiduals}
\alias{ComputeResiduals}
\title{Compute Residuals for Latent Gaussian Count Model}
\usage{
ComputeResiduals(lgc)
}
\arguments{
\item{lgc}{A list object containing the outcome of a model fit.}
}
\value{
A numeric vector of residuals.
}
\description{
Compute Residuals for Latent Gaussian Count Model
}
