% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{dGpois}
\alias{dGpois}
\title{Generalized Poisson PMF}
\usage{
dGpois(y, a, m, log = FALSE)
}
\arguments{
\item{y}{Non-negative integer vector of quantiles.}

\item{a}{Dispersion parameter.}

\item{m}{Mean parameter.}

\item{log}{Logical; if \code{TRUE}, probabilities are returned on the log scale.}
}
\value{
A numeric vector of (log-)probabilities.
}
\description{
Computes the probability mass function (PMF) of the Generalized Poisson distribution
using the mean-parametrization (Famoye 1994, relation 2.4).
}
