% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{pGpois}
\alias{pGpois}
\title{Generalized Poisson CDF (vectorized and scalar)}
\usage{
pGpois(x, a, m)
}
\arguments{
\item{x}{Integer or numeric vector. Values at which to evaluate the CDF.}

\item{a}{Numeric. Dispersion parameter.}

\item{m}{Numeric scalar or vector. Mean parameter(s).}
}
\value{
A numeric vector or matrix of cumulative probabilities.
}
\description{
Computes the cumulative distribution function (CDF) of the Generalized Poisson distribution
for one or more values of the mean parameter \code{m}.
}
\examples{
pGpois(3, 0.4, 2)
}
\seealso{
\code{\link{dGpois}}, \code{\link{qGpois}}
}
